fixed the index page

+++ g.php
@@ -8,12 +8,16 @@
 // //ob_end_clean();
 // echo "<p>latest static page has been generated</p>";
 
-$pages = array("form.php", "list.php", "dock.php", "systemsetting.php", "journal.php", "typography.php", "index.php", "home.php", "dashboard.php", "learning.php", "performance.php");
+$pages = array("form.php", "list.php", "dock.php", "systemsettings.php", "journal.php", "typography.php", "index.php", "home.php", "dashboard.php", "learning.php", "performance.php");
 $arrlength = count($pages);
 
 for ($x = 0; $x < $arrlength; $x++) {
-	generateStaticPage($pages[$x]);
-	echo "<p><strong>$pages[$x]</strong> has been generated</p>";
+	if (file_exists($pages[$x])) {
+		generateStaticPage($pages[$x]);
+		echo "<p><strong>$pages[$x]</strong> has been generated</p>";
+	} else {
+		echo "<p><strong>$pages[$x]</strong> doesn't exist</p>";
+	}
 }
 
 function generateStaticPage($page) {
@@ -24,5 +28,5 @@ function generateStaticPage($page) {
 	$outputfile = str_replace(".php", "", "{$page}.html");
 	file_put_contents($outputfile, ob_get_clean());
 }
-echo "<a href='d.php'>Delet all html files</a>";
+echo "<a href='d.php'>Delete all html files</a>";
 ?>
